WIP transition to package extensions
