+++ Project.toml
@@ -1,6 +1,6 @@
 name = "ADNLPModels"
 uuid = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
-version = "0.7.2"
+version = "0.8.0"
 
 [deps]
 ColPack = "ffa27691-3a59-46ab-a8d4-551f45b8d401"
@@ -11,10 +11,49 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
+[weakdeps]
+Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
+SparseDiffTools = "47a9eef4-7e08-11e9-0b38-333d64bd3804"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
+
+[extensions]
+ADNLPModelsEnzymeExt = "Enzyme"
+ADNLPModelsSparseDiffToolsExt = "SparseDiffTools"
+ADNLPModelsSymbolicsExt = "Symbolics"
+ADNLPModelsZygoteExt = "Zygote"
+
 [compat]
 ColPack = "0.3"
 ForwardDiff = "0.9.0, 0.10.0"
 NLPModels = "0.18, 0.19, 0.20, 0.21"
-Requires = "1"
 ReverseDiff = "1"
-julia = "^1.6"
+julia = "^1.9"
+Enzyme = "0.12"
+SparseDiffTools = "2.3"
+Symbolics = "5.3"
+Zygote = "0.6"
+
+[extras]
+Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
+SparseDiffTools = "47a9eef4-7e08-11e9-0b38-333d64bd3804"
+Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+ManualNLPModels = "30dfa513-9b2f-4fb3-9796-781eabac1617"
+NLPModelsModifiers = "e01155f1-5c6f-4375-a9d8-616dd036575f"
+NLPModelsTest = "7998695d-6960-4d3a-85c4-e1bceb8cd856"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = [
+    "Test",
+    "CUDA",
+    "Enzyme",
+    "SparseDiffTools",
+    "Symbolics",
+    "Zygote",
+    "ManualNLPModels",
+    "NLPModelsModifiers",
+    "NLPModelsTest",
+]
