+++ src/ADNLPModels.jl
@@ -6,7 +6,6 @@ using LinearAlgebra, SparseArrays
 using ColPack, ForwardDiff, ReverseDiff
 # JSO
 using NLPModels
-using Requires
 
 abstract type AbstractADNLPModel{T, S} <: AbstractNLPModel{T, S} end
 abstract type AbstractADNLSModel{T, S} <: AbstractNLSModel{T, S} end
@@ -45,11 +44,25 @@ end
 include("sparse_jacobian.jl")
 include("sparse_hessian.jl")
 
+# Attempt to load some symbols from the Symbolics extensions
+symbolics_ext = Base.get_extension(@__MODULE__, :ADNLPModelsSymbolicsExt)
+if !isnothing(symbolics_ext)
+  SparseSymbolicsADJacobian = symbolics_ext.SparseSymbolicsADJacobian
+  SparseSymbolicsADHessian = symbolics_ext.SparseSymbolicsADHessian
+  SDTSparseADJacobian = symbolics_ext.SDTSparseADJacobian
+
+  # These backends should only be included if the module is loaded
+  predefined_backend[:default][:jacobian_backend] = SparseADJacobian
+  predefined_backend[:default][:jacobian_residual_backend] = SparseADJacobian
+  predefined_backend[:optimized][:jacobian_backend] = SparseADJacobian
+  predefined_backend[:optimized][:jacobian_residual_backend] = SparseADJacobian
+
+  predefined_backend[:default][:hessian_backend] = SparseADHessian
+  predefined_backend[:optimized][:hessian_backend] = SparseReverseADHessian
+end
+
 include("forward.jl")
 include("reverse.jl")
-include("enzyme.jl")
-include("sparse_diff_tools.jl")
-include("zygote.jl")
 include("predefined_backend.jl")
 include("nlp.jl")
 
@@ -178,20 +191,6 @@ function ADNLSModel!(model::AbstractNLSModel; kwargs...)
       model.meta.ucon;
       kwargs...,
     )
-  end
-end
-
-@init begin
-  @require Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7" begin
-    include("sparse_sym.jl")
-
-    predefined_backend[:default][:jacobian_backend] = SparseADJacobian
-    predefined_backend[:default][:jacobian_residual_backend] = SparseADJacobian
-    predefined_backend[:optimized][:jacobian_backend] = SparseADJacobian
-    predefined_backend[:optimized][:jacobian_residual_backend] = SparseADJacobian
-
-    predefined_backend[:default][:hessian_backend] = SparseADHessian
-    predefined_backend[:optimized][:hessian_backend] = SparseReverseADHessian
   end
 end
