+++ test/runtests.jl
@@ -1,22 +1,10 @@
+# Load extension dependencies first
+using SparseDiffTools, Symbolics
+
 using CUDA, LinearAlgebra, SparseArrays, Test
 using ADNLPModels, ManualNLPModels, NLPModels, NLPModelsModifiers, NLPModelsTest
 using ADNLPModels:
   gradient, gradient!, jacobian, hessian, Jprod!, Jtprod!, directional_second_derivative, Hvprod!
-
-@testset "Error without loading package for sparsity pattern" begin
-  f(x) = sum(x)
-  c!(cx, x) = begin
-    cx .= 1
-    return x
-  end
-  nvar, ncon = 2, 1
-  x0 = ones(nvar)
-  cx = rand(ncon)
-  @test_throws ArgumentError ADNLPModels.compute_jacobian_sparsity(c!, cx, x0)
-  @test_throws ArgumentError ADNLPModels.compute_hessian_sparsity(f, nvar, c!, ncon)
-end
-
-using SparseDiffTools, Symbolics
 
 @testset "Test using a NLPModel instead of AD-backend" begin
   include("manual.jl")
